[bdecays] Take into account charm mass scheme uncertainty in B->Vgamma.
The matrix elements of the current-current operators entering B->Vgamma decays (and B->Vll
at low q^2) involves a charm loop, but the renormalization scheme of the charm is not fixed
at the order considered. This leads to a non-negligible uncertainty in B->Vgamma observables.
To account for this fact, instead of using the charm pole mass as before, now a new parameter
'm_c BVgamma' is introduced that allows to fix the charm mass in this quantity by hand.