[lambdab] Parametrization of subleading hadronic effects in Lambdab->Lambdall.
This includes unknown non-factorizable effects like spectator scattering. As in B->Vll and B->Pll decays, subleading effects are parametrized
as polynomials in q2. The central values of the parameters are 0 and their uncertainties correspond to a conservative estimate
of the size of neglected effects. The values chosen correspond to roughly twice the corresponding uncertainties in B->Vll
decays at low q2, and to a comparable uncertainty at high q2.