[classes] Change in the internals of the Constraints class:
Now every 'parameter' (i.e. an actual parameter in the case of the
ParameterConstraints class and an observable in the case of the
Measurement class) can only be subject to a single constraint,
i.e. a single ProbabilityDistribution instance. In fact the
possibility to have several constraints was never used in
practice. In the course of this change, the method `remove_constraints`
was renamed to `remove_constraint`. The old name still works but
will be removed in the future.