[classes] Change to the log_likelihood method of the Constraints class:
This concerns the treatment of parameters (or measurements) that have
correlated uncertainties but shall be ignored in the evaluation of the
likelihood. Previously, this was achieved by setting these parameters to
their central values. While not inconsistent, this can lead to
counterintuitive results when there are two strongly correlated parameters
and one of them is ignored. The new implementation removes the row and
column of the ignored parameter from the covariance matrix of the
multivariate normal constraint, effectively "profiling out" the parameter
(or measurement).

Note that this change only affects fits made using the BayesianFit class,
while fits made using the FastFit class are unaffected.