[fits] New user warning:
If there are several correlated measurements of an observable not
included among the fit observables, this observable will be
"profiled out" in each individual measurement, but the optimal
values can differ from measurement to measurement, which would
be inconsistent. The user will now be warned of such a situation
when instantiating any Fit. The problem can be circumvented by
either

- including the observable in the fit (if physically appropriate),
- ignoring all but one of the correlated measurements (e.g. using
  the exclude_measurements option),
- manually combining the correlated measurements into one,
  such that the profiling is done consistently.