[fits] Major improvement in FastFit:
Previously, the experimental likelihoods had been approximated as Gaussian
by generating a large sample of points and extracting the mean and covariance.
Now, this approximation is dropped: the exact experimental likelihoods are
instead convolved with the theoretical uncertainties (still approximated to
be Gaussian). This allows a consistent treatment of arbitrarily non-Gaussian
experimental likelihoods in arbitrary dimensions in FastFit.