avoid some interpolation artifacts in contour plots (#37)
* avoid some interpolation artifacts in contour plots

If data for which the contours are to be plotted contains very high and
sharp peaks (or other sharp transitions), ringing artifacts appear from
interpolating with splines of order 2 or higher. The ringing might even
yield negative values for purely positive data. This commit avoids these
effects by limiting the z-values to a maximum of twice the largest value
in the levels array and thus effectively cutting off high peaks.

* simplify indexing

* add comment on cutting off large z values in contour plots