[measurements] Remove covariance singularity check.
This check is redundant as it is done at the instantiation of
MultivariateNormalDistribution anyway. Moreover, the latter is
superior as it rescales the covariance by the central values,
making it more stable in case of very disparate scales.

This also removes a "dirty trick" trying to fix singular covariance
matrices ... but this was a dirty trick anyway.