Standardize normalization of KL,KS->l1l2 amplitude
Taking an extra factor of 2 from the prefactor in `br_kll` and putting a factor of sqrt(2) to both short distance `amplitudes` and `amplitudes_LD`. This does not affect the result, of course. Now, the formula for the BR is the more standard one (the same that is used for B->l1 l2 in flavio), and also combining weak-K0-eigenstate amplitudes to those of KL or KS is more intuitive now.