A simple fix to the eps_K prediction (see #172)
As discussed in #172, the RG factors assume the quark masses used
are mq(mq), not mq(160 GeV).
- For the top, we add a new function to the running module get_mt_mt
  which calculates mt(mt) using the RunDec function mOS2mSI and the
  top pole mass.
- For the charm, the input parameter m_c is already mc(mc) from PDG,
  so use that.
- For the up, the input m_u is actually mu(2 GeV), but this is closer
  to mu(mu) than mu(160 GeV), and numerically the up mass is extremely
  unimportant anyway.

The epsK NP test has to be changed, since with the correct value of mc
the charm contribution to imag(M12) is larger (around 30% vs 10% with
the old value), and so the NP doesn't exactly double the SM value.