Add transversity amplitude formalism for Bs -> V ll.
computation of the transversity amplitudes following arxiv:0811.1214 and implementation of the J, h, and s coefficients with the transversity amplitudes.
TODO: add AuxiliaryQuantity behaviour for observables?
TODO: add tests between helicity and transversity amplitudes.

Main functions:
bsvll_obs: drop in replacement for observables_bs "bsvll_obs" for computing the observables.
transversity_amps and transversity_amps_bar to be used as a drop-in replacement for the amplitudes.helicity_amps and amplitudes.helicity_amps_bar used in bsvll_obs to compute the J, h, and s coefficients.